refactor: extract ferrosa-row-bridge crate; ferrosa-postgres drops ferrosa-cql dep (restores D10)
The storage loader's reuse of ferrosa_cql::bridge (for canonical Partition->row
ordering) had re-coupled ferrosa-postgres to the ~54k-LOC ferrosa-cql crate. Fixed
by EXTRACTION, not duplication (duplication would risk silently-divergent row
order, FMEA's top risk).

- New neutral crate ferrosa-row-bridge { codec: encode_value/decode_value + CQL
  type-name parser; row: partition_to_rows[_with_storage_mapping] + key/clustering
  decode }. Deps: ferrosa-common/-sstable/-schema, num-bigint, uuid, tracing — NOT
  ferrosa-cql. Own RowBridgeError.
- ferrosa-cql re-exports the moved functions at their original paths
  (bridge::partition_to_rows_with_storage_mapping / parse_cql_type_in_keyspace,
  types::{encode_value,decode_value}) with From<RowBridgeError> for CqlError, so its
  hundreds of internal callers + full test suite need ZERO changes.
- ferrosa-postgres now depends on ferrosa-row-bridge; ferrosa-cql dep removed.

Decoupling proven: cargo tree -p ferrosa-postgres -e normal | grep ferrosa-cql = 0.
Behavior-preserving: ferrosa-cql suite green (935 lib + integration); ferrosa-postgres
107 lib + 5 live + 2 scram; live differential oracle 26/26 Match (exercises the
extracted decode path vs real Postgres). clippy + fmt clean. Follow-up #8 marked
RESOLVED.
