+++ ferrosa-cql/Cargo.toml
@@ -15,6 +15,7 @@ ferrosa-cluster = { path = "../ferrosa-cluster" }
 ferrosa-common = { path = "../ferrosa-common" }
 ferrosa-net = { path = "../ferrosa-net" }
 ferrosa-index = { path = "../ferrosa-index" }
+ferrosa-row-bridge = { path = "../ferrosa-row-bridge" }
 ferrosa-schema = { path = "../ferrosa-schema" }
 ferrosa-session = { path = "../ferrosa-session" }
 ferrosa-sstable = { path = "../ferrosa-sstable" }
