+++ ferrosa-row-bridge/src/lib.rs
@@ -1,0 +1,71 @@
+//! Neutral storage-row bridge shared by `ferrosa-cql` and `ferrosa-postgres`.
+//!
+//! This crate holds the storage-row decode/codec and `Partition` -> row
+//! decomposition logic that *both* front-ends must agree on byte-for-byte.
+//! Originally these functions lived in `ferrosa-cql` (`bridge.rs` / `types.rs`).
+//! The Postgres front-end reuses the *exact same* decomposition so that its
+//! column ordering matches the CQL read path — duplicating the logic would risk
+//! silently-divergent row ordering (the top FMEA risk for the SQL front-end).
+//!
+//! To let `ferrosa-postgres` reuse it **without** depending on the ~54k-LOC
+//! `ferrosa-cql` crate (decision D10), the closure was extracted here. This
+//! crate depends only on `ferrosa-common`, `ferrosa-sstable`, `ferrosa-schema`,
+//! `num-bigint`, `uuid`, and `tracing` — never on `ferrosa-cql`.
+//!
+//! `ferrosa-cql` re-exports these functions at their original public paths
+//! (`ferrosa_cql::types::{encode_value, decode_value}`,
+//! `ferrosa_cql::bridge::{partition_to_rows_with_storage_mapping, ...}`), so its
+//! internal callers are unaffected.
+//!
+//! ## Modules
+//! - [`codec`] — CQL wire-format `encode_value` / `decode_value` plus the CQL
+//!   type-name parser (`parse_cql_type` / `parse_cql_type_in_keyspace`).
+//! - [`row`] — `Partition` -> row decomposition
+//!   (`partition_to_rows_with_storage_mapping` and friends) and the partition /
+//!   clustering key decoders.
+
+pub mod codec;
+pub mod row;
+
+/// Error returned by the fallible row-bridge functions (`decode_value`,
+/// `parse_cql_type`, `parse_cql_type_in_keyspace`).
+///
+/// The original code returned `ferrosa_cql::error::CqlError::Invalid(String)`;
+/// every failure path in the moved closure used exactly that single variant.
+/// This crate carries a minimal stand-in so it does not depend on `ferrosa-cql`.
+/// `ferrosa-cql` provides `impl From<RowBridgeError> for CqlError` at its
+/// re-export boundary, mapping it back to `CqlError::Invalid` so callers see the
+/// identical error.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct RowBridgeError(pub String);
+
+impl RowBridgeError {
+    /// Construct an invalid-input error with the given message. Mirrors the
+    /// `CqlError::Invalid(msg)` construction in the original code.
+    pub fn invalid(msg: impl Into<String>) -> Self {
+        RowBridgeError(msg.into())
+    }
+
+    /// The error message (the inner `String`).
+    pub fn message(&self) -> &str {
+        &self.0
+    }
+}
+
+impl std::fmt::Display for RowBridgeError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+impl std::error::Error for RowBridgeError {}
+
+pub use codec::{decode_value, encode_value, parse_cql_type, parse_cql_type_in_keyspace};
+pub use row::{
+    decode_clustering, decode_pk, partition_to_rows, partition_to_rows_with_clustering,
+    partition_to_rows_with_storage_mapping, write_partition_raw_rows_with_storage_mapping,
+};
+
+// Liveness helpers are re-exported for `ferrosa-cql`'s remaining metadata
+// decomposition variants, which still live in `ferrosa-cql` but reuse these.
+pub use row::{cell_is_live, ldt_is_expired};
